+++ packages/zosfiles/__tests__/__integration__/cli/hMigrate/ds/__scripts__/migrate_data_set_help.sh
@@ -2,7 +2,7 @@
 set -e
 
 echo "================Z/OS FILES MIGRATE DATA-SET-SEQUENTIAL HELP==============="
-zowe zos-files hMigrate data-set --help
+zowe zos-files migrate data-set --help
 if [ $? -gt 0 ]
 then
     exit $?