+++ packages/zosfiles/__tests__/__integration__/cli/hRecall/ds/__scripts__/command/command_recall_data_set.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+dataSetName=$1
+set -e
+
+echo "================Z/OS FILES RECALL DATA SET==============="
+zowe zos-files hRecall data-set "$dataSetName" 
+if [ $? -gt 0 ]
+then
+    exit $?
+fi