+++ packages/zosfiles/__tests__/__system__/api/methods/hMigrate/HMigrate.system.test.ts
@@ -60,27 +60,27 @@ describe("Migrate Dataset", () => {
                     Imperative.console.info(`Error: ${inspect(err)}`);
                 }
             });
-            // it("should migrate a sequential data set", async () => {
-            //     let error;
-            //     let migrateResponse;
-            //     let listResponse;
+            it("should migrate a sequential data set", async () => {
+                let error;
+                let migrateResponse;
+                let listResponse;
 
-            //     try {
-            //         migrateResponse = await HMigrate.dataSet(REAL_SESSION, dataSet1);
-            //         listResponse = await List.dataSet(REAL_SESSION, dataSet1, listOptions);
-            //         Imperative.console.info(`Response: ${inspect(migrateResponse)}`);
-            //     } catch (err) {
-            //         error = err;
-            //         Imperative.console.info(`Error: ${inspect(err)}`);
-            //     }
+                try {
+                    migrateResponse = await HMigrate.dataSet(REAL_SESSION, dataSet1);
+                    listResponse = await List.dataSet(REAL_SESSION, dataSet1, listOptions);
+                    Imperative.console.info(`Response: ${inspect(migrateResponse)}`);
+                } catch (err) {
+                    error = err;
+                    Imperative.console.info(`Error: ${inspect(err)}`);
+                }
 
-            //     expect(error).toBeFalsy();
+                expect(error).toBeFalsy();
 
-            //     expect(migrateResponse).toBeTruthy();
-            //     expect(migrateResponse.success).toBe(true);
-            //     expect(listResponse.apiResponse.items[0].migr).toBe("YES");
-            //     expect(migrateResponse.commandResponse).toContain(ZosFilesMessages.datasetMigratedSuccessfully.message);
-            // });
+                expect(migrateResponse).toBeTruthy();
+                expect(migrateResponse.success).toBe(true);
+                expect(listResponse.apiResponse.items[0].migr).toBe("YES");
+                expect(migrateResponse.commandResponse).toContain(ZosFilesMessages.datasetMigratedSuccessfully.message);
+            });
             it("should migrate a sequential data set with wait = true", async () => {
                 const migrateOptions: IMigrateOptions = { wait: true };
                 let error;
@@ -111,6 +111,27 @@ describe("Migrate Dataset", () => {
                 } catch (err) {
                     Imperative.console.info(`Error: ${inspect(err)}`);
                 }
+            });
+            it("should migrate a partitioned dataset", async () => {
+                let error;
+                let migrateResponse;
+                let listResponse;
+
+                try {
+                    migrateResponse = await HMigrate.dataSet(REAL_SESSION, dataSet2);
+                    listResponse = await List.dataSet(REAL_SESSION, dataSet2, listOptions);
+                    Imperative.console.info(`Response: ${inspect(migrateResponse)}`);
+                } catch (err) {
+                    error = err;
+                    Imperative.console.info(`Error: ${inspect(err)}`);
+                }
+
+                expect(error).toBeFalsy();
+
+                expect(migrateResponse).toBeTruthy();
+                expect(migrateResponse.success).toBe(true);
+                expect(listResponse.apiResponse.items[0].migr).toBe("YES");
+                expect(migrateResponse.commandResponse).toContain(ZosFilesMessages.datasetMigratedSuccessfully.message);
             });
             it("should migrate a partitioned dataset with wait = true", async () => {
                 const migrateOptions: IMigrateOptions = { wait: true };