+++ packages/zosfiles/__tests__/__system__/cli/hRecall/ds/cli.files.HRecall.ds.system.test.ts
@@ -16,7 +16,7 @@ import { ITestPropertiesSchema } from "../../../../../../../__tests__/__src__/pr
 import { join } from "path";
 import { Session, Imperative } from "@zowe/imperative";
 import { List, Delete, Create, CreateDataSetTypeEnum, IListOptions } from "../../../../..";
-import { IMigrateOptions } from "../../../../../src/api/methods/hMigrate/doc/IMigrateOptions";
+import { IRecallOptions } from "../../../../../src/api/methods/hRecall/doc/IRecallOptions";
 import { inspect } from "util";
 
 let TEST_ENVIRONMENT: ITestEnvironment;
@@ -30,22 +30,22 @@ let REAL_SESSION: Session;
 const listOptions: IListOptions = { attributes: true };
 
 const scriptsLocation = join(__dirname, "__scripts__", "command");
-const migrateScript = join(scriptsLocation, "command_migrate_data_set.sh");
-const migrateScriptWait = join(scriptsLocation, "command_migrate_data_set_wait.sh");
+const recallScript = join(scriptsLocation, "command_recall_data_set.sh");
+const recallScriptWait = join(scriptsLocation, "command_recall_data_set_wait.sh");
 
-describe("Migrate Dataset", () => {
+describe("Recall Dataset", () => {
   beforeAll(async () => {
     TEST_ENVIRONMENT = await TestEnvironment.setUp({
       tempProfileTypes: ["zosmf"],
-      testName: "zos_migrate_data_set"
+      testName: "zos_recall_data_set"
     });
     defaultSystem = TEST_ENVIRONMENT.systemTestProperties;
     REAL_SESSION = TestEnvironment.createZosmfSession(TEST_ENVIRONMENT);
 
     user = defaultSystem.zosmf.user.trim().toUpperCase();
-    dataSetName1 = `${user}.SDATAC.MIGR`;
-    dataSetName2 = `${user}.PDATAC.MIGR`;
-    dataSetName3 = `${user}.FAIL.MIGR`;
+    dataSetName1 = `${user}.SDATAC.REC`;
+    dataSetName2 = `${user}.PDATAC.REC`;
+    dataSetName3 = `${user}.FAIL.REC`;
   });
 
   afterAll(async () => {
@@ -72,26 +72,26 @@ describe("Migrate Dataset", () => {
           Imperative.console.info(`Error: ${inspect(err)}`);
         }
       });
-      it("Should migrate a data set", async () => {
-        const response = runCliScript(migrateScript, TEST_ENVIRONMENT, [dataSetName1]);
+      it("Should recall a data set", async () => {
+        const response = runCliScript(recallScript, TEST_ENVIRONMENT, [dataSetName1]);
         const list1 = await List.dataSet(REAL_SESSION, dataSetName1, listOptions);
 
-        expect(list1.apiResponse.items[0].migr).toBe("YES");
+        expect(list1.apiResponse.items[0].migr).toBe("NO");
 
         expect(response.stderr.toString()).toBe("");
         expect(response.status).toBe(0);
-        expect(response.stdout.toString()).toContain("Data set migraton requested.");
+        expect(response.stdout.toString()).toContain("Data set recall requested.");
       });
-      it("Should migrate a data set with wait = true", async () => {
-        const migrateOptions: IMigrateOptions = { wait: true };
-        const response = runCliScript(migrateScriptWait, TEST_ENVIRONMENT, [dataSetName1, migrateOptions]);
+      it("Should recall a data set with wait = true", async () => {
+        const recallOptions: IRecallOptions = { wait: true };
+        const response = runCliScript(recallScriptWait, TEST_ENVIRONMENT, [dataSetName1, recallOptions]);
         const list1 = await List.dataSet(REAL_SESSION, dataSetName1, listOptions);
 
-        expect(list1.apiResponse.items[0].migr).toBe("YES");
+        expect(list1.apiResponse.items[0].migr).toBe("NO");
 
         expect(response.stderr.toString()).toBe("");
         expect(response.status).toBe(0);
-        expect(response.stdout.toString()).toContain("Data set migraton requested.");
+        expect(response.stdout.toString()).toContain("Data set recall requested.");
       });
     });
     describe("Partitioned Data Set", () => {
@@ -102,26 +102,26 @@ describe("Migrate Dataset", () => {
           Imperative.console.info(`Error: ${inspect(err)}`);
         }
       });
-      it("Should migrate a data set", async () => {
-        const response = runCliScript(migrateScript, TEST_ENVIRONMENT, [dataSetName2]);
+      it("Should recall a data set", async () => {
+        const response = runCliScript(recallScript, TEST_ENVIRONMENT, [dataSetName2]);
         const list2 = await List.dataSet(REAL_SESSION, dataSetName2, listOptions);
 
-        expect(list2.apiResponse.items[0].migr).toBe("YES");
+        expect(list2.apiResponse.items[0].migr).toBe("NO");
 
         expect(response.stderr.toString()).toBe("");
         expect(response.status).toBe(0);
-        expect(response.stdout.toString()).toContain("Data set migraton requested.");
+        expect(response.stdout.toString()).toContain("Data set recall requested.");
       });
-      it("Should migrate a data set with wait = true", async () => {
-        const migrateOptions: IMigrateOptions = { wait: true };
-        const response = runCliScript(migrateScriptWait, TEST_ENVIRONMENT, [dataSetName2, migrateOptions]);
+      it("Should recall a data set with wait = true", async () => {
+        const recallOptions: IRecallOptions = { wait: true };
+        const response = runCliScript(recallScriptWait, TEST_ENVIRONMENT, [dataSetName2, recallOptions]);
         const list2 = await List.dataSet(REAL_SESSION, dataSetName2, listOptions);
 
-        expect(list2.apiResponse.items[0].migr).toBe("YES");
+        expect(list2.apiResponse.items[0].migr).toBe("NO");
 
         expect(response.stderr.toString()).toBe("");
         expect(response.status).toBe(0);
-        expect(response.stdout.toString()).toContain("Data set migraton requested.");
+        expect(response.stdout.toString()).toContain("Data set recall requested.");
       });
     });
   });
@@ -135,11 +135,11 @@ describe("Migrate Dataset", () => {
         }
       });
       it("Should throw an error if a missing data set name is selected", async () => {
-        const response = runCliScript(migrateScript, TEST_ENVIRONMENT, ["MISSING.DATA.SET", dataSetName3]);
+        const response = runCliScript(recallScript, TEST_ENVIRONMENT, ["MISSING.DATA.SET", dataSetName3]);
 
         expect(response.stderr.toString()).toBeTruthy();
         expect(response.status).toBe(1);
-        expect(response.stdout.toString()).not.toContain("Data set migraton requested.");
+        expect(response.stdout.toString()).not.toContain("Data set recall requested.");
       });
     });
   });